+++ src/components/domain/dictionary/WordsList/index.tsx
@@ -1,6 +1,6 @@
+import WordListItem from '@/components/shared/WordListItem'
+import { words } from '@/components/shared/WordListItem/data'
 import { FilterType, SimpleWordType } from '@/types/word'
-import { words } from '../OneWordItem/data'
-import OneWordItem from '../OneWordItem'
 
 type WordsListProps = {
   category: FilterType
@@ -10,7 +10,7 @@ export default function WordsList({ category }: WordsListProps) {
   return (
     <>
       {words.map((word: SimpleWordType, idx: number) => (
-        <OneWordItem key={word.id} word={word} />
+        <WordListItem key={word.id} word={word} />
       ))}
     </>
   )
