+++ src/components/shared/WordListItem/data.ts
